requirements.txt
Django==3.2.9
reportlab==3.6.2
djangorestframework==3.12.4
gunicorn==20.1.0
drf-pdf==0.2.0